+++ onboarding-enabler-java/src/main/java/org/zowe/apiml/eurekaservice/client/util/EurekaInstanceConfigCreator.java
@@ -15,18 +15,23 @@ import org.zowe.apiml.config.ApiInfo;
 import org.zowe.apiml.eurekaservice.client.config.*;
 import org.zowe.apiml.exception.MetadataValidationException;
 import org.zowe.apiml.exception.ServiceDefinitionException;
+import org.zowe.apiml.util.MapUtils;
 import org.zowe.apiml.util.UrlUtils;
 
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import static org.zowe.apiml.constants.EurekaMetadataDefinition.*;
 
 public class EurekaInstanceConfigCreator {
+
+    private final MapUtils mapUtils;
+
+    public EurekaInstanceConfigCreator(MapUtils mapUtils) {
+        this.mapUtils = mapUtils;
+    }
 
     public EurekaInstanceConfig createEurekaInstanceConfig(ApiMediationServiceConfig config) throws ServiceDefinitionException {
         ApimlEurekaInstanceConfig result = new ApimlEurekaInstanceConfig();
@@ -130,52 +135,7 @@ public class EurekaInstanceConfigCreator {
     }
 
     public Map<String, String> flattenMetadata(Map<String, Object> configurationMetadata) {
-        return flattenMap(null, configurationMetadata);
+        return mapUtils.flattenMap(null, configurationMetadata);
     }
 
-    private Map<String, String> flattenMap(String rootKey, Map<String, Object> collection) {
-        if (collection == null || collection.isEmpty()) {
-            return Collections.emptyMap();
-        }
-
-        Map<String, String> result = new HashMap<>();
-
-        for (Map.Entry<String, Object> entry : collection.entrySet()) {
-
-            if (entry.getValue() == null) {
-                result.put( mergeKey(rootKey, entry.getKey()), "");
-                continue;
-            }
-
-
-            if (entry.getValue() instanceof Map) {
-                result.putAll(flattenMap(mergeKey(rootKey, entry.getKey()), (Map<String, Object>)entry.getValue()));
-                continue;
-            }
-
-            if (entry.getValue() instanceof String ||
-                entry.getValue() instanceof Boolean ||
-                entry.getValue() instanceof Integer ||
-                entry.getValue() instanceof Double ||
-                entry.getValue() instanceof Float) {
-                result.put(mergeKey(rootKey, entry.getKey()), entry.getValue().toString());
-                continue;
-            }
-
-            if (entry.getValue() instanceof List) {
-                throw new IllegalArgumentException("List parsing is not supported");
-            }
-            if (entry.getValue().getClass().isArray()) {
-                throw new IllegalArgumentException("Array parsing is not supported");
-            }
-            throw new IllegalArgumentException(String.format("Cannot parse key: %s with value %s", entry.getKey(), entry.getValue().toString()));
-        }
-
-        return result;
-    }
-
-
-    private String mergeKey(String rootKey, String newKey) {
-        return rootKey != null ? rootKey + "." + newKey : newKey;
-    }
 }
